removing unused styles,removing framer motion variants from FlightsListDetailedInformation and moving it to its own file for better readability
