+++ src/App.module.scss
@@ -115,7 +115,6 @@
             padding: 1px;
             border-radius: 50px;
             overflow: hidden;
-            // background: hsla(221, 93%, 22%, 1);
             background: rgb(16, 38, 68);
             background: linear-gradient(
               90deg,
@@ -124,12 +123,6 @@
             );
             padding: 2.5px;
             border: 0;
-
-            // background: radial-gradient(
-            //   circle,
-            //   hsla(221, 100%, 91%, 1) 30%,
-            //   hsla(221, 93%, 22%, 1) 100%
-            // );
 
             .search_bar {
               width: 100%;
