+++ src/components/FlightsListDetailedData/FlightsListDetailedData.tsx
@@ -5,6 +5,7 @@ import { useContext, useEffect, useState } from "react";
 import { FetchContext } from "../../context/fetch-context";
 import Dropdown from "../Dropdown/Dropdown";
 import { AnimatePresence, motion } from "framer-motion";
+import variants from "./framer_variants";
 
 interface FlightsListDetailedDataProps {
   el: FlightDataType;
@@ -33,93 +34,6 @@ const FlightsListDetailedData = (props: FlightsListDetailedDataProps) => {
     departureData,
     selectedPage,
   ]);
-
-  const variants = {
-    parent_container: {
-      scale: 1,
-    },
-    initial: {
-      backgroundColor: "#ffffff",
-      borderRadius: "80px 80px 80px 80px",
-      color: "#000000",
-      opacity: 1,
-      transition: {
-        duration: 0.5,
-        delay: 0.5,
-      },
-    },
-    active: {
-      backgroundColor: "#1f3d60",
-      borderRadius: "80px 80px 0px 0px",
-      color: "#ffffff",
-    },
-    hover_active: {
-      outline: "rgba(31, 61, 96, 0)",
-      scale: 1,
-      transition: {
-        duration: 0.2,
-        ease: "easeInOut",
-      },
-    },
-    hover_inactive: {
-      outline: "2px solid rgba(31, 61, 96, 1)",
-      transition: {
-        duration: 0.1,
-        ease: "easeInOut",
-      },
-    },
-    codeInitial: {
-      color: "#1f3d60",
-      // Set the initial color of the spans
-      transition: {
-        duration: 0.2,
-        ease: "easeInOut",
-        delay: 0.3,
-      },
-    },
-    codeActive: {
-      color: "#ffffff",
-      transition: {
-        duration: 0.2,
-        ease: "easeInOut",
-        delay: 0.3,
-      }, // Set the active color of the spans
-    },
-    arrowInitial: {
-      top: "100%",
-      left: "50%",
-      x: "-50%",
-      y: "-50%",
-      backgroundColor: "#ffffff",
-      rotate: 0,
-      borderRadius: "80px 80px 80px 80px",
-    },
-    arrowActive: {
-      top: "100%",
-      left: "50%",
-      x: "-50%",
-      y: "-50%",
-      backgroundColor: "#1f3d60",
-      borderRadius: "80px",
-      rotate: "180deg",
-    },
-    arrowHoverActive: {
-      borderRadius: "80px",
-      y: "-40%",
-      transition: {
-        delay: 0,
-        duration: 0,
-      },
-    },
-    arrowHoverInactive: {
-      borderRadius: "80px",
-      y: "-40%",
-      transition: {
-        delay: 0,
-        duration: 0,
-      },
-    },
-  };
 
   return (
     <motion.div
